reogranize start and game
