+++ web/src/EthersContext.js
@@ -1,0 +1,15 @@
+import React, { useState } from 'react';
+
+const EthersContext = React.createContext([{}, () => {}]);
+
+const EthersContextProvider = (props) => {
+  const [state, setState] = useState({});
+
+  return (
+    <EthersContext.Provider value={[state, setState]}>
+      {props.children}
+    </EthersContext.Provider>
+  );
+}
+
+export { EthersContext, EthersContextProvider };
